+++ addons/cartel/deprecated/CfgGroups_Deprecated.hpp
@@ -1,0 +1,150 @@
+class TACU_Cartel_O_G_Soldado_Infantry {
+    name = "Soldados";
+
+    class TACU_Cartel_G_O_Soldado_Patrol {
+        name = "Patrol";
+        side = 0;
+        faction = "TACU_Cartel_O";
+        icon = "\a3\ui_f\data\map\markers\nato\o_inf.paa";
+        class unit0 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_01";
+            side = 0;
+            rank = "SERGEANT";
+            position[] = {0, 0, 0};
+        };
+        class unit1 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_02";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {5, -5, 0};
+        };
+    };
+
+    class TACU_Cartel_G_O_Soldado_Fireteam {
+        name = "Soldado Fireteam";
+        side = 0;
+        faction = "TACU_Cartel_O";
+        icon = "\a3\ui_f\data\map\markers\nato\o_inf.paa";
+        class unit0 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_01";
+            side = 0;
+            rank = "SERGEANT";
+            position[] = {0, 0, 0};
+        };
+        class unit1 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_04";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {5, -5, 0};
+        };
+        class unit2 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_AT";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {-5, -5, 0};
+        };
+        class unit3 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_MG";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {10, -10, 0};
+        };
+    };
+
+    class TACU_Cartel_G_O_Soldado_Squad {
+        name = "Soldados Squad";
+        side = 0;
+        faction = "TACU_Cartel_O";
+        icon = "\a3\ui_f\data\map\markers\nato\o_inf.paa";
+        class unit0 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_01";
+            side = 0;
+            rank = "SERGEANT";
+            position[] = {0, 0, 0};
+        };
+        class unit1 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_02";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {5, -5, 0};
+        };
+        class unit2 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_03";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {-5, -5, 0};
+        };
+        class unit3 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_04";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {10, -10, 0};
+        };
+        class unit4 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_AT";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {-10, -10, 0};
+        };
+        class unit5 {
+            vehicle = "TACU_Cartel_U_O_Soldado_Rifleman_MG";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {15, -15, 0};
+        };
+    };
+};
+
+class TACU_Cartel_G_O_Sicario_Infantry {
+    name = "Sicarios";
+
+    class TACU_Cartel_G_O_Sicario_Patrol {
+        name = "Sicario Patrol";
+        side = 0;
+        faction = "TACU_Cartel_O";
+        icon = "\a3\ui_f\data\map\markers\nato\o_inf.paa";
+        class unit0 {
+            vehicle = "TACU_Cartel_U_O_Sicario_Rifleman_03";
+            side = 0;
+            rank = "SERGEANT";
+            position[] = {0, 0, 0};
+        };
+        class unit1 {
+            vehicle = "TACU_Cartel_U_O_Sicario_Rifleman_04";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {5, -5, 0};
+        };
+    };
+
+    class TACU_Cartel_G_Sicario_Fireteam {
+        name = "Sicario Fireteam";
+        side = 0;
+        faction = "TACU_Cartel_O";
+        icon = "\a3\ui_f\data\map\markers\nato\o_inf.paa";
+        class unit0 {
+            vehicle = "TACU_Cartel_U_O_Sicario_Rifleman_03";
+            side = 0;
+            rank = "SERGEANT";
+            position[] = {0, 0, 0};
+        };
+        class unit1 {
+            vehicle = "TACU_Cartel_U_O_Sicario_Rifleman_04";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {5, -5, 0};
+        };
+        class unit2 {
+            vehicle = "TACU_Cartel_U_O_Sicario_Rifleman_AT";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {-5, -5, 0};
+        };
+        class unit3 {
+            vehicle = "TACU_Cartel_U_O_Sicario_Rifleman_MG";
+            side = 0;
+            rank = "PRIVATE";
+            position[] = {10, -10, 0};
+        };
+    };
+};
